+++ app/Exports/template.php
@@ -2,7 +2,6 @@
 
 namespace App\Exports;
 
-use App\Models\products;
 use Maatwebsite\Excel\Concerns\FromView;
 use Illuminate\Contracts\View\View;
 
@@ -14,7 +13,7 @@ class template implements FromView
         $this -> products = $productLists;
     }
     public function view(): View
-    {  
+    {
         if(!empty($this->products)){
             // dd($this->products);
             return view('main.products.exportTemplate', [
@@ -23,7 +22,7 @@ class template implements FromView
         }else{
             return view('Không có dữ liệu');
         }
-        
+
     }
 
 }
