+++ app/Models/Orders.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class Orders extends Model
+{
+    use HasFactory;
+
+    protected $fillable = [
+        'URLImages',
+        'nameProduct',
+        'quatity',
+        'priceProduct',
+        'purchaseProduct',
+        'description',
+        'created_at',
+        'updated_at'
+    ];
+
+    /**
+     * The attributes that should be hidden for serialization.
+     *
+     * @var array<int, string>
+     */
+    protected $hidden = [
+
+    ];
+
+    /**
+     * The attributes that should be cast.
+     *
+     * @var array<string, string>
+     */
+    protected $casts = [
+
+    ];
+}
