+++ app/Models/Orders_repository.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\DB;
+
+class Orders_repository extends Model
+{
+    use HasFactory;
+
+    // insert order
+    public function insertOrders($data)
+    {
+        DB::table('orders')->insert([$data]);
+    }
+
+    // get order by ID transactions
+    public function getOrderByIDTransaction($id)
+    {
+        return DB::table('orders')
+                ->select('*')
+                ->where("ID_Transactions", "=", $id)
+                ->get();
+    }
+}
