+++ resources/views/clients/blocks/dashboard.blade.php
@@ -1,3 +1,3 @@
 <div style="margin-top: 70px; height: 600px; text-align:center">
-    <img src="{{ asset('assets/clients/images/page.jpg') }}" alt="Wellcome"  width="700px" height="500px">
+    <img src="{{ asset('assets/clients/images/page.png') }}" alt="Wellcome"  width="700px" height="500px">
 </div>
